+++ client/src/model.rs
@@ -1,1 +1,2 @@
-pub mod edit_item;
+pub mod item_page;
+pub mod common;
